refactor: extract action select into its own component
